Resolve remaining recursive constraint quality findings
